Created new layout for results, moved much code there (UNTESTED)
