+++ app/src/main/res/values/strings.xml
@@ -22,4 +22,7 @@
     <string name="help">Help</string>
     <string name="e_searchButtonString">Search</string>
     <string name="e_searchBarHint">Enter search term</string>
+    <string name="e_welcomeText">Welcome to the Ticket Master Search</string>
+    <string name="e_citySearchBarHint">Enter city</string>
+    <string name="e_radiusSearchBarHint">Enter Radius</string>
 </resources>
